SqlManager can now assign values from a given row into a respected type.
This is a heavy WIP. However at the moment, this can be described as follows:
After loading all rows, each one is being tokenized, with a delim character of "|".
Then, values are getting converted to a respected type (if needed) and inserted in
a constructor of a desired type (Student, School, etc.). Finally new object is being
inserted into a specified map, with a ID as a key and object instance as a value.
Some additional handling is being done in the mean-time, like adding a student to a
vector inside a school object.

At the moment the functionality is only inside one class and the logic is not
separated (a.k.a. not so SOLID!). This will be changed during the later development.