Added initial schema creation in the sqlManager.
Created `attribute` struct and `table` class to ensure modularity
and customizability during the schema creation process. Each table
contains it's schema and will eventually contain all of the possible entries
in raw format. `Table` class will be able to `makeFormula` in order to return
a `CREATE TABLE` formula just from one line call.

SqlManager will not have an open Database from the initalization but rather
after `openDb` method is called. That way a valid db file is no longer required
to create an SqlManager object.

WsManager will now create initial schema file and a starting database file.

FileManager can now return a pointer to an opened file if one exists.
In addition it can also parse a given file an return a value as a vector.
(Not sure if will last - for now it's just a not used feature)