Moved the SQL and DB intialization logic into WsManager.
SqlManager is now more of a abstract class that can be reused in different projects.
Generally the initial creation of a schema file is now a reponsibility of a wsManager.
SqlManager will only load it whenever a program is being launched.

Additional changes:
- Moved tokenizer from a class to a common utility source file.
- Table will now use unordered_map to store attributes. That way
it will store the id in the same location every time.
- 2 separate DB callback functions were created for the code reusability.