Added supprot for updating entries.
If a misstake was done, or a something needs to be updated a new method was
introduced. Update will take the original object and new object with new parameters.
The update works as follows: SqlAdapter gets the new object, and checkes it values
if a missmatch occurs, the new value is inserted inside the `New Value Packet`.
When there is at least 1 packet to change it will call the SQL database to
alter the existing entry.