Turned core types into more abstract structs.
From now on, each core type is a derived of an "Entry" struct.
This decouple a lot inside the sqlAdapter class.
From now on each type is being treated the same, and adapter
does not need to know which type it have to deal with, that
responsibility is now only inside the session class.
Now only 3 to 4 functions will be used to translate into SQL.