Added `Grade` and `CourseSubjectWeight` to the core types.
Some changes are going to be delivered as new handles will be introduced
and some changes will be made for existing handles.

`School` will now use a shared pointer to a school instead of a copy.
(Finally!)
`InputHandler` will use a reserved bit composition to determine if a valid
input was provided or not.